WHOOHOO! New Version of Geometry Fake!
In the new version we have updated the old levels, and added additional secrets and prepared the game to version 1.3. Game Rate from 1 to 5, so we can understand that the game is wrong and what needs to be done better. Write comments, leave your feedback. I hope you like what we do for you!